+++ sample-android/src/main/java/com/github/panpf/zoomimage/sample/ui/examples/compose/GlideZoomAsyncImageSample.kt
@@ -1,6 +1,5 @@
 package com.github.panpf.zoomimage.sample.ui.examples.compose
 
-import android.widget.Toast
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
@@ -8,9 +7,11 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.tooling.preview.Preview
 import com.github.panpf.sketch.fetch.newResourceUri
+import com.github.panpf.tools4a.toast.ktx.showShortToast
 import com.github.panpf.zoomimage.GlideZoomAsyncImage
 import com.github.panpf.zoomimage.compose.glide.internal.ExperimentalGlideComposeApi
 import com.github.panpf.zoomimage.sample.R
+import com.github.panpf.zoomimage.sample.ui.util.compose.toShortString
 import com.github.panpf.zoomimage.sample.util.sketchUri2GlideModel
 
 @OptIn(ExperimentalGlideComposeApi::class)
@@ -32,10 +33,10 @@ fun GlideZoomAsyncImageSample(sketchImageUri: String) {
             state = state,
             scrollBar = scrollBar,
             onTap = {
-                Toast.makeText(context, "Click", Toast.LENGTH_SHORT).show()
+                context.showShortToast("Click (${it.toShortString()})")
             },
             onLongPress = {
-                Toast.makeText(context, "Long click", Toast.LENGTH_SHORT).show()
+                context.showShortToast("Long click (${it.toShortString()})")
             }
         )
     }