+++ sample-android/src/main/java/com/github/panpf/zoomimage/sample/ui/examples/compose/SketchZoomAsyncImageSample.kt
@@ -1,6 +1,5 @@
 package com.github.panpf.zoomimage.sample.ui.examples.compose
 
-import android.widget.Toast
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -8,8 +7,10 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.tooling.preview.Preview
 import com.github.panpf.sketch.fetch.newResourceUri
 import com.github.panpf.sketch.request.DisplayRequest
+import com.github.panpf.tools4a.toast.ktx.showShortToast
 import com.github.panpf.zoomimage.SketchZoomAsyncImage
 import com.github.panpf.zoomimage.sample.R
+import com.github.panpf.zoomimage.sample.ui.util.compose.toShortString
 
 @Composable
 fun SketchZoomAsyncImageSample(sketchImageUri: String) {
@@ -30,10 +31,10 @@ fun SketchZoomAsyncImageSample(sketchImageUri: String) {
             state = state,
             scrollBar = scrollBar,
             onTap = {
-                Toast.makeText(context, "Click", Toast.LENGTH_SHORT).show()
+                context.showShortToast("Click (${it.toShortString()})")
             },
             onLongPress = {
-                Toast.makeText(context, "Long click", Toast.LENGTH_SHORT).show()
+                context.showShortToast("Long click (${it.toShortString()})")
             }
         )
     }